requirements.txt
requests
pyyaml
python-dotenv
mcrcon
